rename BlogAuthor to ContentAuthor (#6353)
* rename BlogAuthor to ContentAuthor

* .

* sorted

* sorted

* migrations

* migrations

* migrations
