+++ network-api/networkapi/wagtailpages/pagemodels/blog/blog.py
@@ -12,9 +12,7 @@ from wagtail.admin.edit_handlers import (
 from wagtail.core import blocks
 from wagtail.core.models import Orderable, Page
 from wagtail.core.fields import StreamField
-from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.snippets.edit_handlers import SnippetChooserPanel
-from wagtail.snippets.models import register_snippet
 
 from taggit.models import TaggedItemBase
 from modelcluster.fields import ParentalKey, ParentalManyToManyField
@@ -31,6 +29,7 @@ from ...utils import (
 
 from .blog_category import BlogPageCategory
 from .blog_index import BlogIndexPage
+from ..content_author import ContentAuthor
 
 base_fields = [
     ('paragraph', blocks.RichTextBlock(
@@ -57,32 +56,12 @@ class BlogPageTag(TaggedItemBase):
     content_object = ParentalKey('wagtailpages.BlogPage', on_delete=models.CASCADE, related_name='tagged_items')
 
 
-@register_snippet
-class BlogAuthor(models.Model):
-
-    name = models.CharField(max_length=70, blank=False)
-    image = models.ForeignKey(
-        'wagtailimages.Image',
-        on_delete=models.SET_NULL,
-        blank=True,
-        null=True,
-    )
-
-    panels = [
-        FieldPanel("name"),
-        ImageChooserPanel("image"),
-    ]
-
-    def __str__(self):
-        return self.name
-
-
 class BlogAuthors(Orderable):
     """This allows us to select one or more blog authors from Snippets."""
 
     page = ParentalKey("wagtailpages.BlogPage", related_name="authors")
     author = models.ForeignKey(
-        BlogAuthor,
+        ContentAuthor,
         on_delete=models.CASCADE,
     )
